ESIF LF: Fix Critical Policy Not Taking Action at Boot
BUG: None in JIRA. 5317484 in HSD Client SW

This fix addresses a corner case which is when critical
shutdown condition is met at system boot, the DPTF Java
system service is not up fast enough to receive the
event, therefore failing to shut down the system.

The fix will repeatedly retry sending the uevent, so
that when the DPTF Java service is finally up it will
receive the uevent and shut down the system accordingly.

This patch also diables sending sleep uevent
as the team is discovering more corner cases that
require further testing, and since the customers for this
release do not enable the DPTF sleep feature, it was
decided that the DPTF sleep is disabled for the
Baytrail-M Android Beta release.

Signed-off-by: Brian Bian <brian.bian@intel.com>

Change-Id: I84d59eb9e68ba0ab6d7b5938562bcd2965b64f5e
Reviewed-on: https://android.intel.com/199641
Reviewed-by: Bian, Brian <brian.bian@intel.com>
Tested-by: sys_pccgand <sys_pccgand@intel.com>
Reviewed-by: Deuskar, Prafulla <prafulla.deuskar@intel.com>
Reviewed-by: Wiles, AmyX L <amyx.l.wiles@intel.com>
Tested-by: Wiles, AmyX L <amyx.l.wiles@intel.com>
Reviewed-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>
Tested-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>